Organizing stuff preparing for cli mode
